+++ src/math/math.rs
@@ -1,11 +1,11 @@
 // convert degress to radians
-pub fn as_radians(angle: f64) -> f64 {
-    angle * std::f64::consts::PI / 180.0
+pub fn as_radians(angle: f32) -> f32 {
+    angle * std::f32::consts::PI / 180.0
 }
 
 // convert radians to degrees
-pub fn as_degrees(angle: f64) -> f64 {
-    angle * 180.0 / std::f64::consts::PI
+pub fn as_degrees(angle: f32) -> f32 {
+    angle * 180.0 / std::f32::consts::PI
 }
 
 
@@ -19,11 +19,11 @@ mod test {
 
     #[test]
     fn test_as_radians() {
-        assert_eq!(as_radians(180.0), std::f64::consts::PI);
+        assert_eq!(as_radians(180.0), std::f32::consts::PI);
     }
 
     #[test]
     fn test_as_degrees() {
-        assert_eq!(as_degrees(std::f64::consts::PI), 180.0);
+        assert_eq!(as_degrees(std::f32::consts::PI), 180.0);
     }
 }
